Make sure task fields estimate, estimatedHours, xp and payout are floats on output

+++ app/Adapters/Task.php
@@ -33,10 +33,10 @@ class Task implements AdaptersInterface
             $this->task->{$key} = $value;
         }
 
-        $this->task->estimate = sprintf('%.2f', $this->task->estimatedHours);
-        $this->task->estimatedHours = $originalEstimate;
-        $this->task->xp = sprintf('%.2f', $this->task->xp);
-        $this->task->payout = sprintf('%.2f', $mappedValues['payout']);
+        $this->task->estimate = (float) sprintf('%.2f', $this->task->estimatedHours);
+        $this->task->estimatedHours = (float) $originalEstimate;
+        $this->task->xp = (float) sprintf('%.2f', $this->task->xp);
+        $this->task->payout = (float) sprintf('%.2f', $mappedValues['payout']);
 
         $taskStatus = $profilePerformance->perTask($this->task);
